server/requirements.txt
astroid==1.6.5
autopep8==1.3.5
certifi==2018.4.16
chardet==3.0.4
click==6.7
coreapi==2.3.0
coreapi-cli==1.0.9
coreschema==0.0.4
Django==1.11.2
django-cors-headers==2.0.2
django-debug-toolbar==1.8
django-timelog==0.4
djangorestframework==3.6.3
idna==2.6
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
Markdown==2.6.8
MarkupSafe==1.0
mccabe==0.6.1
olefile==0.45.1
progressbar==2.5
pycodestyle==2.3.1
Pygments==2.2.0
pylint==1.9.2
pytz==2018.4
requests==2.18.4
six==1.11.0
sqlparse==0.2.4
texttable==1.4.0
uritemplate==3.0.0
urllib3==1.22
uWSGI==2.0.15
wrapt==1.10.11
